src/sample-code/clickup-cdk/test/widget.test.ts
import { core } from '@time-loop/cdk-library';
import { App, assertions } from 'aws-cdk-lib';
import { Namer } from 'multi-convention-namer';
import { WidgetStack } from './../src/widget';

const commonProps = {
  businessUnit: core.BusinessUnit.PRODUCT,
  clickUpEnvironment: core.ClickUpEnvironment.PRODUCTION,
  clickUpRole: core.ClickUpRole.APP,
  confidentiality: core.Confidentiality.PUBLIC,
  namedEnv: core.Environment.usDev('us-west-2'),
};

describe('Widget', () => {
  describe('default', () => {
    const app = new App();
    const stack = new WidgetStack(app, new Namer(['test']), {
      ...commonProps,
      managedPolicyName: 'fakeName',
    });

    const template = assertions.Template.fromStack(stack);
    test('snapshot', () => {
      expect(template.toJSON()).toMatchSnapshot();
    });

    test('managedPolicyName should work', () => {
      template.hasResourceProperties('AWS::IAM::ManagedPolicy', {
        ManagedPolicyName: 'fakeName',
      });
    });

    test('policy should reference key', () => {
      template.hasResourceProperties('AWS::IAM::ManagedPolicy', {
        PolicyDocument: {
          Statement: assertions.Match.arrayWith([
            assertions.Match.objectLike({
              Resource: {
                'Fn::GetAtt': [
                  assertions.Match.anyValue(), // TODO: figure out how to actually reference the generated Kms key
                  'Arn',
                ],
              },
            }),
          ]),
        },
      });
    });
  });
});
